+++ src/main/java/com/kl/grooveo/boundedContext/member/controller/MemberController.java
@@ -31,8 +31,8 @@ import com.kl.grooveo.boundedContext.library.service.SoundTrackService;
 import com.kl.grooveo.boundedContext.member.dto.FindPasswordFormDTO;
 import com.kl.grooveo.boundedContext.member.dto.FindUsernameFormDTO;
 import com.kl.grooveo.boundedContext.member.dto.JoinFormDTO;
+import com.kl.grooveo.boundedContext.member.dto.ModifyEmailFormDTO;
 import com.kl.grooveo.boundedContext.member.entity.Member;
-import com.kl.grooveo.boundedContext.member.form.ModifyEmailForm;
 import com.kl.grooveo.boundedContext.member.form.ModifyNickNameForm;
 import com.kl.grooveo.boundedContext.member.form.ModifyPasswordForm;
 import com.kl.grooveo.boundedContext.member.service.MemberService;
@@ -210,8 +210,8 @@ public class MemberController {
 
 	@PreAuthorize("isAuthenticated()")
 	@PostMapping("/myPage/modifyEmail")
-	public String modifyEmail(@Valid ModifyEmailForm modifyEmailForm) {
-		RsData<Member> member = memberService.modifyEmail(rq.getMember(), modifyEmailForm.getEmail());
+	public String modifyEmail(@Valid ModifyEmailFormDTO modifyEmailFormDTO) {
+		RsData<Member> member = memberService.modifyEmail(rq.getMember(), modifyEmailFormDTO.getEmail());
 
 		if (member.isFail()) {
 			return rq.historyBack(member);
