+++ src/main/java/com/kl/grooveo/boundedContext/member/dto/ModifyEmailFormDTO.java
@@ -1,18 +1,17 @@
-package com.kl.grooveo.boundedContext.member.form;
+package com.kl.grooveo.boundedContext.member.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
-@AllArgsConstructor
-public class ModifyEmailForm {
+@Setter
+public class ModifyEmailFormDTO {
 
 	@NotBlank
 	@Email
 	@Size(min = 3, max = 30)
 	private String email;
 }
-
